+++ elastic-job-lite/elastic-job-lite-spring/src/test/java/com/dangdang/ddframe/AllLiteSpringTests.java
@@ -23,8 +23,6 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({
-        AllSpringJobTests.class,
-    })
+@SuiteClasses(AllSpringJobTests.class)
 public class AllLiteSpringTests {
 }